refactor: split resumeCoordinator into runStateVerifier + resumeQueueRecovery (finding 130)
Extract state-check helpers and types from 816-line resumeCoordinator.ts into:
- runStateVerifier.ts: ResumeAnalysis/DiagnosticSeverity/ResumeDiagnostic/ResumeOptions types
  and checkRunStatus, checkPendingApprovals, checkIntegrity, analyzeLastError,
  checkQueueState, generateRecommendations helpers
- resumeQueueRecovery.ts: receives checkQueueFiles (moved from coordinator)
- resumeCoordinator.ts: reduced to 294-line thin orchestrator with re-exports for backward compat
